requirements.txt
whisper==1.0
pyannote.audio==2.1.1
yt-dlp==2023.01.06
torch>=1.12.0
torchaudio>=0.12.0
